refactor(charts): extract config serialization into ChartConfigMixin
Move to_config, from_config, _serialize_annotation, and _rebuild_annotations
out of the Chart base class into a new _chart_config.ChartConfigMixin, mixed
back in as the first base so MRO resolves them unchanged. Behavior preserving;
visual baselines untouched.
